backend/api/pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "local-vault-backend"
version = "0.1.0"
description = "Local-first secure password manager backend"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.115.6,<1.0.0",
  "uvicorn[standard]>=0.32.1,<1.0.0",
  "pydantic>=2.10.3,<3.0.0",
  "pydantic-settings>=2.7.0,<3.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "alembic>=1.14.0,<2.0.0",
  "argon2-cffi>=23.1.0,<24.0.0",
  "cryptography>=43.0.3,<44.0.0",
  "python-multipart>=0.0.17,<0.1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4,<9.0.0",
  "httpx>=0.28.1,<1.0.0",
  "hypothesis>=6.122.3,<7.0.0",
  "bandit>=1.7.10,<2.0.0",
  "pip-audit>=2.8.0,<3.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
